refactor: move code to updaters
